WIP on simplified RDFa helpers
The new helpers are ill-documented.  However, they base themselves on
the same principles with the change that arguments provide you with a
way of controlling the output.  As such, there are only three helpers
to take into account: with-rdfa-context (which takes a model and sets
up the necessary context), ctx.get (which lets you link to a related
resource, or render a specific property) and ctx.each (which lists
over a has-many relationship).  Documentation can be greatly improved
and we suspect small changes in the interface are still in order.