Initial .drone.yml file
Initial commit to start the drone configuration. In the future, we would
want repos to use Ember 3.28 LTS, and therefore this drone set-up uses
NodeJS 16 LTS instead of NodeJS 14 LTS, as this is the most recent
Docker image avaliable. This is why a minor change in package.json was
necessary.